+++ test/sub_scope_manager_blue_print_test.dart
@@ -26,7 +26,7 @@ void main() {
           // that produces the row height of the row
           var i = 0;
           for (final rowScope in rowScopes) {
-            final rowHeightNode = rowScope.findNode<int>('RowHeight')!;
+            final rowHeightNode = rowScope.findNode<int>('rowHeight')!;
             expect(rowHeightNode.produce([rowHeights], 0), rowHeights[i]);
             i++;
           }