+++ test/sub_scope_manager_test.dart
@@ -9,14 +9,14 @@ import 'package:supply_chain/supply_chain.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('SubScopeManager', () {
+  group('subScopeManager', () {
     group('example', () {
       test('should work', () {
         // Get some variables
         final subScopeManager = SubScopeManager.example();
         final scope = subScopeManager.scope;
         final scm = scope.scm;
-        final rowHeightsNode = scope.findNode<List<int>>('RowHeights')!;
+        final rowHeightsNode = scope.findNode<List<int>>('rowHeights')!;
         expect(rowHeightsNode, isNotNull);
 
         // Init suppliers
@@ -37,14 +37,14 @@ void main() {
 
         // Now we should have 3 row scopes
         expect(scope.children.length, 3);
-        final row0 = scope.child('Row0')!;
-        final row1 = scope.child('Row1')!;
-        final row2 = scope.child('Row2')!;
+        final row0 = scope.child('row0')!;
+        final row1 = scope.child('row1')!;
+        final row2 = scope.child('row2')!;
 
         // Each row scope should have the right height
-        expect(row0.node<int>(key: 'RowHeight')?.product, 10);
-        expect(row1.node<int>(key: 'RowHeight')?.product, 20);
-        expect(row2.node<int>(key: 'RowHeight')?.product, 30);
+        expect(row0.node<int>(key: 'rowHeight')?.product, 10);
+        expect(row1.node<int>(key: 'rowHeight')?.product, 20);
+        expect(row2.node<int>(key: 'rowHeight')?.product, 30);
 
         // Update the row heights
         rowHeightsNode.product = [40, 50, 60, 70];
@@ -56,11 +56,11 @@ void main() {
         expect(scm.nodes, hasLength(6));
 
         // The row scopes should have the new heights
-        final row3 = scope.child('Row3')!;
-        expect(row0.node<int>(key: 'RowHeight')?.product, 40);
-        expect(row1.node<int>(key: 'RowHeight')?.product, 50);
-        expect(row2.node<int>(key: 'RowHeight')?.product, 60);
-        expect(row3.node<int>(key: 'RowHeight')?.product, 70);
+        final row3 = scope.child('row3')!;
+        expect(row0.node<int>(key: 'rowHeight')?.product, 40);
+        expect(row1.node<int>(key: 'rowHeight')?.product, 50);
+        expect(row2.node<int>(key: 'rowHeight')?.product, 60);
+        expect(row3.node<int>(key: 'rowHeight')?.product, 70);
 
         // Remove row heights
         rowHeightsNode.product = [10];
@@ -68,7 +68,7 @@ void main() {
         scm.testFlushTasks(); // Note: One flush should be enough. Check this.
         expect(scm.nodes, hasLength(3));
         expect(scope.children.length, 1);
-        expect(row0.node<int>(key: 'RowHeight')?.product, 10);
+        expect(row0.node<int>(key: 'rowHeight')?.product, 10);
       });
     });
 
@@ -79,8 +79,8 @@ void main() {
       var producedScopeBluePrints = <ScopeBluePrint>[];
 
       when(() => bluePrint.initialProduct).thenReturn([]);
-      when(() => bluePrint.key).thenReturn('SubScopeManager');
-      when(() => bluePrint.suppliers).thenReturn(['RowHeights']);
+      when(() => bluePrint.key).thenReturn('subScopeManager');
+      when(() => bluePrint.suppliers).thenReturn(['rowHeights']);
       when(() => bluePrint.produce).thenReturn(
         (List<dynamic> components, List<ScopeBluePrint> previous) =>
             producedScopeBluePrints,
@@ -102,10 +102,10 @@ void main() {
       // Assume that the produce method adds a child scope
       producedScopeBluePrints = [
         ScopeBluePrint(
-          key: 'Row0',
+          key: 'row0',
           nodes: [
             NodeBluePrint<int>(
-              key: 'RowHeight',
+              key: 'rowHeight',
               initialProduct: 10,
               suppliers: [],
               produce: (components, previousProduct) => 10,
@@ -121,16 +121,16 @@ void main() {
       // Change the produced scope. It should add one additional node.
       producedScopeBluePrints = [
         ScopeBluePrint(
-          key: 'Row0',
+          key: 'row0',
           nodes: [
             NodeBluePrint<int>(
-              key: 'RowHeight',
+              key: 'rowHeight',
               initialProduct: 10,
               suppliers: [],
               produce: (components, previousProduct) => 10,
             ),
             NodeBluePrint<int>(
-              key: 'RowHeight2',
+              key: 'rowHeight2',
               initialProduct: 20,
               suppliers: [],
               produce: (components, previousProduct) => 20,
@@ -146,10 +146,10 @@ void main() {
       // Change the produced scope. It should remove the last node.
       producedScopeBluePrints = [
         ScopeBluePrint(
-          key: 'Row0',
+          key: 'row0',
           nodes: [
             NodeBluePrint<int>(
-              key: 'RowHeight',
+              key: 'rowHeight',
               initialProduct: 10,
               suppliers: [],
               produce: (components, previousProduct) => 10,